combine add and edit category in one screen
