+++ app/src/main/java/com/example/tudeeapp/presentation/screen/categories/CategoriesScreen.kt
@@ -43,7 +43,7 @@ fun CategoriesScreen(viewModel: CategoriesViewModel = koinViewModel()) {
         state = state,
         onClickCategory = {navController.navigate(Screens.CategoryDetails(it))},
         onClickAddCategory = {
-            navController.navigate(Screens.AddCategoryScreen)
+            navController.navigate(Screens.CategoryFormScreen())
         }
         )
 }
