+++ app/src/main/java/com/example/tudeeapp/presentation/screen/categoriesForm/CategoryFormViewModel.kt
@@ -1,14 +1,14 @@
 package com.example.tudeeapp.presentation.screen.categoriesForm
 
 import android.net.Uri
+import android.util.Log
 import androidx.core.net.toUri
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.navigation.toRoute
 import com.example.tudeeapp.domain.TaskServices
+import com.example.tudeeapp.domain.models.Category
 import com.example.tudeeapp.presentation.mapper.toResDrawables
-import com.example.tudeeapp.presentation.navigation.Screens
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -16,16 +16,25 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
-class CategoryFormViewModel(val taskServices: TaskServices, savedStateHandle: SavedStateHandle) : ViewModel() {
+class CategoryFormViewModel(
+    val taskServices: TaskServices,
+    savedStateHandle: SavedStateHandle
+) : ViewModel() {
 
     private val _state = MutableStateFlow(CategoryFormState())
     val state: StateFlow<CategoryFormState> = _state.asStateFlow()
 
-    val args = savedStateHandle.toRoute<Screens.CategoryFormEditScreen>()
-
     init {
+        val args = savedStateHandle.get<Long>("categoryId")
+        Log.d("EditCategory", "Received categoryId: $args")
+        if (args != null) {
+            loadCategory(args)
+        }
+    }
+
+    private fun loadCategory(id: Long) {
         viewModelScope.launch {
-            val category = taskServices.getCategoryById(args.categoryId).first()
+            val category = taskServices.getCategoryById(id).first()
             _state.update {
                 it.copy(
                     categoryName = category.title,
@@ -41,35 +50,65 @@ class CategoryFormViewModel(val taskServices: TaskServices, savedStateHandle: Sa
         }
     }
 
-    fun updateCategoryName(newCategoryName: String) {
-        _state.update { it.copy(categoryName = newCategoryName) }
+    fun updateCategoryName(newName: String) {
+        _state.update { it.copy(categoryName = newName) }
     }
 
     fun updateImage(uri: Uri) {
         _state.update { it.copy(imageUri = uri) }
     }
 
+    fun submitCategory() {
+        if (state.value.categoryId != 0L) {
+            editCategory()
+        } else {
+            addCategory()
+        }
+    }
 
     fun editCategory() {
         viewModelScope.launch {
-            taskServices.editCategory(
-                id = state.value.categoryId,
-                title = state.value.categoryName,
-                imageUrl = getImageUrl(state.value.imageUri.toString())
-            )
+            try {
+                taskServices.editCategory(
+                    id = state.value.categoryId,
+                    title = state.value.categoryName,
+                    imageUrl = getImageUrl(state.value.imageUri.toString())
+                )
+                _state.update { it.copy(successMessage = "Category updated successfully") }
+            } catch (e: Exception) {
+                _state.update { it.copy(errorMessage = "Failed to edit category") }
+            }
         }
     }
 
-
-    private fun getImageUrl(url: String): String{
-        if (url.startsWith("content")){
+    private fun getImageUrl(url: String): String {
+        if (url.startsWith("content")) {
             return url
-        }
-        else{
+        } else {
             return url.getLastPartAfterSlash()
         }
-
     }
+
+    private fun addCategory() {
+        viewModelScope.launch {
+            try {
+                val category = Category(
+                    title = state.value.categoryName,
+                    imageUrl = state.value.imageUri.toString(),
+                    tasksCount = 0,
+                    isPredefined = false
+                )
+                taskServices.addCategory(category)
+                _state.update { it.copy(successMessage = "Category added successfully") }
+            } catch (e: Exception) {
+                _state.update {
+                    it.copy(errorMessage = e.message ?: "Unexpected error")
+                }
+            }
+        }
+    }
+
+
 }
 
 fun String.getLastPartAfterSlash(): String {
